test: better align `ssr-compiler` fixture tests with `engine-server` (#4406)
