+++ packages/@lwc/engine-server/src/__tests__/fixtures/attribute-boolean-expr/index.js
@@ -1,2 +1,3 @@
 export const tagName = 'x-attribute-boolean-expr';
 export { default } from 'x/attribute-boolean-expr';
+export * from 'x/attribute-boolean-expr';
