+++ packages/@lwc/engine-server/src/__tests__/fixtures/attribute-namespace/index.js
@@ -1,2 +1,3 @@
 export const tagName = 'x-attribute-namespace';
 export { default } from 'x/attribute-namespace';
+export * from 'x/attribute-namespace';
