+++ packages/@lwc/engine-server/src/__tests__/fixtures/attribute-scoped-id/index.js
@@ -1,2 +1,3 @@
 export const tagName = 'x-attribute-scoped-id';
 export { default } from 'x/scoped-id';
+export * from 'x/scoped-id';
